agd/requirements-test.txt
pytest
pytest-cov
numpy
Pillow
torch
tqdm
